refactor: export movieService as an object
